feat: add github workflow to build brandi
Still gotta add some way to download the built binary.
Sadly, the binary that this will build won't be statically built, and will not
work on most non-debian systems. I have added a docker step as well, which
should build a docker image based on alpine (where brandi won't run), but it
keep building a gigantic >700MB image probably based on ubuntu. I am too much in
flow right now to dig into nix docs, so I am gonna let it slide.

To use the binary built with this CI step, it should be run on an ubuntu system