feat: empty spa using react
Since we are building a dashboard, I think it is appropriate to build it as a
single page application. I am choosing React that's what I have most experience
with. I already have plenty of new stuff to learn concurrently; so I am choosing
not to go with Vue although I very much want to.